+++ src/schemathesis/cli/commands/run/__init__.py
@@ -284,6 +284,12 @@ DEFAULT_PHASES = ["examples", "coverage", "fuzzing", "stateful"]
     is_eager=True,
 )
 @grouped_option(
+    "--report-ndjson-path",
+    help="Custom path for NDJSON events file",
+    type=click.File("w", encoding="utf-8"),
+    is_eager=True,
+)
+@grouped_option(
     "--report-preserve-bytes",
     help="Retain exact byte sequence of payloads in cassettes, encoded as base64",
     type=bool,
@@ -467,6 +473,7 @@ def run(
     report_junit_path: LazyFile | None = None,
     report_vcr_path: LazyFile | None = None,
     report_har_path: LazyFile | None = None,
+    report_ndjson_path: LazyFile | None = None,
     report_preserve_bytes: bool | None = None,
     output_sanitize: bool | None = None,
     output_truncate: bool | None = None,
@@ -526,6 +533,7 @@ def run(
         junit_path=report_junit_path.name if report_junit_path else None,
         vcr_path=report_vcr_path.name if report_vcr_path else None,
         har_path=report_har_path.name if report_har_path else None,
+        ndjson_path=report_ndjson_path.name if report_ndjson_path else None,
         directory=Path(report_directory),
         preserve_bytes=report_preserve_bytes,
     )